Create Dependencies type, and move functions onto it
* Move stack functions into stacks
* Move version line functions into version_lines
* Rename BuildpackToml to BuildpackTOML and Sha to SHA, per Go style
guide
[#169324300]
