+++ tasks/update-cnb-dependency/buildpack_toml.go
@@ -8,7 +8,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-type BuildpackToml struct {
+type BuildpackTOML struct {
 	API       string
 	Buildpack struct {
 		ID      string
@@ -16,32 +16,28 @@ type BuildpackToml struct {
 		Version string
 	} `toml:"buildpack"`
 	Metadata Metadata
-	Order    []Order
-	Stacks   []Stack
+	Orders   Orders `toml:"order"`
+	Stacks   Stacks `toml:"stack"`
 }
 
 type Metadata struct {
 	IncludeFiles               []string `toml:"include_files"`
-	Dependencies               []Dependency
-	DependencyDeprecationDates []DependencyDeprecationDate `toml:"dependency_deprecation_dates"`
+	Dependencies               Dependencies
+	DependencyDeprecationDates DeprecationDates `toml:"dependency_deprecation_dates"`
 }
 
-type Stack struct {
-	ID string
+func (buildpackTOML BuildpackTOML) LoadExpandedDependencies() Dependencies {
+	return buildpackTOML.Metadata.Dependencies.ExpandByStack()
 }
 
-func (buildpackToml BuildpackToml) LoadExpandedDependencies() []Dependency {
-	return expandDependenciesForEachStack(buildpackToml.Metadata.Dependencies)
-}
-
-func (buildpackToml BuildpackToml) WriteToFile(filepath string) error {
-	buildpackTomlFile, err := os.OpenFile(filepath, os.O_RDWR|os.O_CREATE, 0666)
+func (buildpackTOML BuildpackTOML) WriteToFile(filepath string) error {
+	buildpackTOMLFile, err := os.OpenFile(filepath, os.O_RDWR|os.O_CREATE, 0666)
 	if err != nil {
 		return errors.Wrap(err, fmt.Sprintf("failed to open buildpack.toml at: %s", filepath))
 	}
-	defer buildpackTomlFile.Close()
+	defer buildpackTOMLFile.Close()
 
-	if err := toml.NewEncoder(buildpackTomlFile).Encode(buildpackToml); err != nil {
+	if err := toml.NewEncoder(buildpackTOMLFile).Encode(buildpackTOML); err != nil {
 		return errors.Wrap(err, "failed to save updated buildpack.toml")
 	}
 	return nil
