+++ tasks/update-cnb-dependency/dependencies.go
@@ -4,26 +4,80 @@ import (
 	"errors"
 	"fmt"
 	"path/filepath"
-	"regexp"
+	"sort"
 
 	"github.com/blang/semver"
 )
 
+type Dependencies []Dependency
+
 type Dependency struct {
 	ID           string
 	Name         string `toml:",omitempty"`
-	Sha256       string
+	SHA256       string
 	Source       string `toml:",omitempty"`
-	SourceSha256 string `toml:"source_sha256,omitempty"`
+	SourceSHA256 string `toml:"source_sha256,omitempty"`
 	Stacks       []string
 	URI          string
 	Version      string
 }
 
-const AnyStack = "any-stack"
-const TinyStack = "tiny"
+func (deps Dependencies) MergeDependencyLists(newDeps Dependencies) (Dependencies, error) {
+	depsMap := map[string]Dependency{}
 
-func sortDependencies(deps []Dependency) func(i, j int) bool {
+	for _, dep := range deps {
+		depsMap[makeKey(dep)] = dep
+	}
+	for _, dep := range newDeps {
+		depsMap[makeKey(dep)] = dep
+	}
+
+	allDeps := Dependencies{}
+	for _, dep := range depsMap {
+		allDeps = append(allDeps, dep)
+	}
+
+	sort.Slice(allDeps, allDeps.sortDependencies())
+	return allDeps, nil
+}
+
+func (deps Dependencies) RemoveOldDeps(depID, versionLine string, keepN int) (Dependencies, error) {
+	if keepN <= 0 {
+		return nil, errors.New("please specify a valid number of versions (>0) to retain")
+	}
+
+	retainedDeps := Dependencies{}
+	retainedPerStack := map[string]int{}
+
+	versionLineConstraint, err := getVersionLineConstraint(versionLine)
+	if err != nil {
+		return nil, err
+	}
+
+	for i := len(deps) - 1; i >= 0; i-- {
+		dep := deps[i]
+		depVersion, err := semver.Parse(dep.Version)
+		if err != nil {
+			return nil, err
+		}
+
+		differentDep := dep.ID != depID
+		differentVersionLine := !versionLineConstraint(depVersion)
+		haveNotRetainedNForStack := retainedPerStack[dep.Stacks[0]] < keepN
+
+		if differentDep || differentVersionLine {
+			retainedDeps = append(retainedDeps, dep)
+		} else if haveNotRetainedNForStack {
+			retainedDeps = append(retainedDeps, dep)
+			retainedPerStack[dep.Stacks[0]]++
+		}
+	}
+
+	sort.Slice(retainedDeps, retainedDeps.sortDependencies())
+	return retainedDeps, nil
+}
+
+func (deps Dependencies) sortDependencies() func(i, j int) bool {
 	return func(i, j int) bool {
 		if deps[i].ID != deps[j].ID {
 			return deps[i].ID < deps[j].ID
@@ -40,9 +94,8 @@ func sortDependencies(deps []Dependency) func(i, j int) bool {
 	}
 }
 
-func expandDependenciesForEachStack(deps []Dependency) []Dependency {
-	var expandedDeps []Dependency
-
+func (deps Dependencies) ExpandByStack() Dependencies {
+	var expandedDeps Dependencies
 	for _, dep := range deps {
 		if len(dep.Stacks) == 1 {
 			expandedDeps = append(expandedDeps, dep)
@@ -58,8 +111,8 @@ func expandDependenciesForEachStack(deps []Dependency) []Dependency {
 	return expandedDeps
 }
 
-func loadDependenciesFromBinaryBuildsForDep(dep Dependency, depOrchestratorConfig DependencyOrchestratorConfig) ([]Dependency, error) {
-	var depsToAdd []Dependency
+func loadDependenciesFromBinaryBuildsForDep(dep Dependency, depOrchestratorConfig DependencyOrchestratorConfig) (Dependencies, error) {
+	var depsToAdd Dependencies
 
 	buildMetadataPaths, err := filepath.Glob(filepath.Join("builds", "binary-builds-new", dep.ID, fmt.Sprintf("%s-*.json", dep.Version)))
 	if err != nil {
@@ -76,7 +129,23 @@ func loadDependenciesFromBinaryBuildsForDep(dep Dependency, depOrchestratorConfi
 	return depsToAdd, nil
 }
 
-func constructDependenciesFromBuildMetadata(dep Dependency, buildMetadataPath string, depOrchestratorConfig DependencyOrchestratorConfig) ([]Dependency, error) {
+func (deps Dependencies) containsDependency(dep Dependency) bool {
+	_, exists := deps.findDependency(dep)
+	return exists
+}
+
+func (deps Dependencies) findDependency(dep Dependency) (Dependency, bool) {
+	for _, d := range deps {
+		if d.ID == dep.ID && d.Version == dep.Version && d.Stacks[0] == dep.Stacks[0] {
+			return d, true
+		}
+	}
+	return Dependency{}, false
+}
+
+func makeKey(dep Dependency) string { return dep.ID + dep.Version + dep.Stacks[0] }
+
+func constructDependenciesFromBuildMetadata(dep Dependency, buildMetadataPath string, depOrchestratorConfig DependencyOrchestratorConfig) (Dependencies, error) {
 	var buildMetadata BuildMetadata
 	if err := loadJSON(buildMetadataPath, &buildMetadata); err != nil {
 		return nil, err
@@ -87,65 +156,20 @@ func constructDependenciesFromBuildMetadata(dep Dependency, buildMetadataPath st
 		return nil, err
 	}
 
-	var deps []Dependency
+	var deps Dependencies
 	for _, stack := range stacks {
 		deps = append(deps, Dependency{
 			ID:           dep.ID,
 			Name:         depOrchestratorConfig.V3DepNames[dep.ID],
-			Sha256:       buildMetadata.Sha256,
+			SHA256:       buildMetadata.SHA256,
 			Source:       buildMetadata.Source.URL,
-			SourceSha256: buildMetadata.Source.Sha256,
+			SourceSHA256: buildMetadata.Source.SHA256,
 			Stacks:       []string{stack},
 			URI:          buildMetadata.URL,
 			Version:      dep.Version,
 		})
 	}
 	return deps, nil
-}
-
-func determineStacks(buildMetadataPath string, dep Dependency, depOrchestratorConfig DependencyOrchestratorConfig) ([]string, error) {
-	stackRegexp := regexp.MustCompile(`\/(?:\.|\d)*-(.*)\.json$`)
-	matches := stackRegexp.FindStringSubmatch(buildMetadataPath)
-	if len(matches) != 2 {
-		return nil, errors.New(fmt.Sprintf("expected to find one stack name in filename (%s) but found: %v", filepath.Base(buildMetadataPath), matches[1:]))
-	}
-	stack := matches[1]
-
-	if stack == AnyStack {
-		return handleAnyStack(dep, depOrchestratorConfig)
-	} else if stackIsDeprecated(stack, depOrchestratorConfig.DeprecatedStacks) {
-		return nil, nil
-	}
-
-	for stackName, stackID := range depOrchestratorConfig.V3Stacks {
-		if stack == stackName {
-			return []string{stackID}, nil
-		}
-	}
-	return nil, errors.New(fmt.Sprintf("%s is not a valid stack", stack))
-}
-
-func handleAnyStack(dep Dependency, config DependencyOrchestratorConfig) ([]string, error) {
-	var stacks []string
-	for stack, stackID := range config.V3Stacks {
-		if stack == TinyStack && !includeTiny(dep.ID, config.IncludeTiny) {
-			continue
-		}
-		stacks = append(stacks, stackID)
-	}
-
-	if len(stacks) == 0 {
-		return nil, errors.New("stack is 'any-stack' but no stacks are configured, check dependency-builds.yml")
-	}
-	return stacks, nil
-}
-
-func stackIsDeprecated(stack string, deprecatedStacks []string) bool {
-	return arrayContains(stack, deprecatedStacks)
-}
-
-func includeTiny(id string, includeTinyStacks []string) bool {
-	return arrayContains(id, includeTinyStacks)
 }
 
 func arrayContains(item string, array []string) bool {
