+++ tasks/update-cnb-dependency/orders.go
@@ -1,5 +1,7 @@
 package main
 
+type Orders []Order
+
 type Order struct {
 	Group []Group
 }
@@ -9,7 +11,7 @@ type Group struct {
 	Version string
 }
 
-func UpdateOrderDependencyVersion(orders []Order, dep Dependency) []Order {
+func (orders Orders) UpdateOrderDependencyVersion(dep Dependency) Orders {
 	for i, order := range orders {
 		for j, group := range order.Group {
 			if group.ID == dep.ID {
